+++ src/game/hudmessage.h
@@ -7,13 +7,13 @@ bool enabled;
 int starttime,endtime;
 float x,y;
 string text;
-int r,g,b;
+Uint8 r,g,b;
 hudmessage();
  // posx,posy - position,(start/stop)delay - delay before the message shown and  the message will disappear after being shown in milliseconds
 };
 namespace game
 {
-extern int hudmessagecolors[3];
+extern Uint8 hudmessagecolors[3];
 extern hudmessage hmessages[MAXHUDCHANNELS];
 
 extern void clearhudmessages();
